Update FAQ: systemctl usage
When a service is not active, `systemctl status` return a non-zero exit code, thus `systemctl status abrtd && systemctl status abrt-ccpp` would possibly show only the (inactive) status of `abrtd`, and nothing about `abrt-ccpp`.
Passing all service names to a single invocation of `systemctl status` show the status for both.

Similarly, we can restart both services in a single call, and that must be done as root (or use sudo). `sudo` makes it easier to copy-paste the snippet.